Fix MultipleClientPasswordsSyncTest.Sanity
This test had some issues with accidentally nested RunLoops.  The
attached bug includes a small essay that thoroughly describes the issue.

The solution implemented here is to prevent nesting by not allowing
multiple calls to IsExitConditionRequired() to occur on the stack at the
same time.  It also ensures that the requests to check the exit
condition are not fully ignored; it will retry the check if a call was
attempted and rejected while the current one was in progress.

This CL also disables self-notifications for the
MultipleClientPasswordsSyncTest.  Since the test no longer depends on
AwaitQuiescence, self-notifications should be no longer necessary.

BUG=363247

Review URL: https://codereview.chromium.org/250943005