sync: Include control types in node browser
Explicitly add the control types to the list of types to fetch nodes for
during a request to GetAllNodes().  The control types don't register
with the ProfileSyncService at startup like the other types do, so
they're not in the list of GetRegisteredDirectoryDataTypes like the
others.

BUG=368012

Review URL: https://codereview.chromium.org/251133005