cc: Fix PictureLayerImpl::SanityCheck condition
PictureLayerImpl::ReleaseResources calls
set_needs_update_draw_properties, expecting as a result that if it is
false again, that CalculateContentsScale is called in between. However
this assumption is not true as CalculateDrawProperties can skip layers.

Instead, also check should_update_tile_priorities_ which is set in
CalculateContentsScale.

BUG=358350

Review URL: https://codereview.chromium.org/258053011