Add Shill to ONC translation of IPConfig.
So far, the ONC IPConfig object was only validated but not translated to or from Shill.
This adds the one direction, from Shill to ONC.

This allows exposing the current IP config state of Shill through ONC, e.g. with in the networkingPrivate API.

BUG=288288

Review URL: https://codereview.chromium.org/249193003