cc: Add support for partial swaps when using impl-side painting by only damaging viewport area covered by newly initialized tiles.
In the current implementation, during frame draw 
and visible tile update, LTHI was setting damage 
to the entire viewport area, irrespective of the 
initialized tiled content rect.
As part of the patch we are trying to ascertain newly 
initialized tile rect from TileManager, notify LTHI about it, 
and save it as damage rect in LayerImpl.
During frame draw, when the damage for surfaces and 
layers are tracked, we query for these damaged layer rects 
and unify with the DamageTracker's damage rect.

BUG=

Review URL: https://codereview.chromium.org/219963005