Fix scale on axis may be negative.
This patch modifies judgment condition whether is equal to zero for double type, avoid unnecessary sqrt calls in more case.
If "a" and "b" are both equal to zero, and "c" is less then zero, currently ScaleOnAxis will return "c" which is negative. So, we should ensure that the result is greater than zero.

BUG=none.

Review URL: https://codereview.chromium.org/253083002