Ensure stale values are not used in Mac App Launcher layout.
Cocoa's NSCollectionView has a buggy layout if it is first shown while
its parent view is scrolling. We work around this by correcting the
layout. However, if the launcher window is not visible the corrected
value is not queried. This results in a stale layout being used when
subsequently shown.

This CL ensures stale values are replaced when correcting the layout.

It also ensures an item move triggered outside the app list updates
layout values. Previously this had an optimization to do a simpler
rearrangement, but it would skip over the corrections.

BUG=364495
TEST=With the mac app launcher hidden, uninstall an app on the first
page of apps from chrome://extensions. Show the lanucher - layout should
be correct.

Review URL: https://codereview.chromium.org/261493003