New policies: enable/disable relay; port range
Add new Chromoting policies:

"RemoteAccessHostAllowRelayedConnection"
bool, default true
It can be set to false to disable the use of relay servers if NAT
traversal is enabled.

"RemoteAccessHostUdpPortRange"
string, default ""
Can be used to specify a range in the form "<min-port>-<max-port>" to
restrict the range of UDP ports available to the host for connections.
E.g. "12400-12409"

BUG=355168

Review URL: https://codereview.chromium.org/209323002