cc: Change required_for_activation bookkeeping.
This patch changes the bookkeeping of whether required for activation
tiles all received memory during AssignGpuMemory. The new approach
doesn't rely on the fact that we process every tile in that loop, thus
allowing us to early out or use another approach at getting tiles that
does not guarantee that every tile will be visited.

The end goal is to evaluate if we can activate when the rasterizer
notifies us that all tasks given to it that are required for
activation are finished rather than trying to figure this out
when we schedule tasks.

R=reveman, enne

Review URL: https://codereview.chromium.org/257773009