sync: Add NonBlockingDataTypeManager
The new class manages the set of NonBlockingDataTypeControllers.  Its
logic could have fit into the ProfileSyncService, but the code is a bit
cleaner if we factor it out into a separate class.

This CL does actually instantiate and make use of the
NonBlockingDataTypeManager, so it is not entirely a no-op.  However,
without any registered non-blocking types, there will be no
NonBlockingDataTypeControllers for it to manage, so its functionality
will be limited.

BUG=351005

Review URL: https://codereview.chromium.org/251143003