GN help and operator fixes.
This fixes error reporting for binary operators, which would previously lose
the error if one of the sides of the operators had a failure. The result would
be that it would give a misleading error about adding "None".

Also fixes references to "to_build_path" to refer to "rebase_path".

R=ajwong@chromium.org

Review URL: https://codereview.chromium.org/254183002