Add implicit hard deps to GN.
Automatically deduce the hard dependency flag from the target type. Code targets never have side effects that require the hard dep flag, and actions almost always do.

This additionally forces all hard deps of all recursive dependencies of a target to be order-only deps of a target. This solves the problem of indirectly depending on an action that generates a source file and depending on the results of that action. Previously (and in GYP) this would result in cases where the first build is wrong.

To avoid an explosion of hard deps on each source file, this adds an input dependency stamp file to coalesce these hard deps, so each source needs only one order-only dependency. This stamp file is written if there are any order-only dependencies, even in cases where there is only one and we might be better off skipping this and writing it directly. The additional performance from doing this didn't seem worth adding the additional complexity for the optimization.

TBR=scottmg@chromium.org
Original review: https://codereview.chromium.org/259353002

Review URL: https://codereview.chromium.org/269463002