Pepper: Move nmf downloading out of trusted plugin.
This change moves NMF downloading for non-data URLs outside the trusted plugin
(data URLs were moved in a previous change). This is another step to removing
all dependencies on FileDownloader in the trusted plugin.

It's difficult to reuse the URLLoaderHost code from components/nacl (since
URLLoaderHost is not in content/public), so this change uses WebURLLoader
directly.

TBR=dglazkov
BUG=239656

Review URL: https://codereview.chromium.org/251383005