sync: Expose DirectoryDebugInfoEmitters in engine
Changes the lifetime of DirectoryTypeDebugInfoEmitters,
DirectoryCommitContributors, and DirectoryUpdateHandlers.  Under the old
lifetime management design, we had no state that lived beyond
a configure cycle.  This would be bad, since we want the debug info
counter values to remain after the initial configure is done.  The
refactoring ensures that the DirectoryTypeDebugInfoEmitters will live as
long as their data type is enabled.  As a side effect, the commit
contributors and update handlers get to live that long, too.

Plumbs the DirectoryTypeDebugInfoEmitters through the commit contributor
and update handler classes.  This prepares them for updating counters
and emitting events from sync/engine/ code, although this functionality
won't be implemented until the next patch in the series.

BUG=328606

Review URL: https://codereview.chromium.org/260613002