End touch editing on mouse and keyboard events
Whenever a mouse or keyboard event happens anywhere in any display,
touch text seleciton should be deactivated. We do not want to deactivate
if mouse event is due to a touch action, therefore we check if mouse
events are not disabled (because of a touch action).

BUG=286030

Review URL: https://codereview.chromium.org/257573003