Allow use of EGL in browser tests with --use-gl=egl.
Browser tests on some ARM devices need to use EGL, but our browser tests
do not currently allow the use of --use-gl and they default to GLX.

Allow the use of --use-gl, but have browser tests override it with the
--override-use-gl-with-osmesa-for-tests flag by default, this ensures
that --use-gl on its own does not change the behaviour of the tests.

The --use-gpu-in-tests flag allows the --use-gl flag to take effect, as
the browser tests will try to use hardware gpu then.

R=piman
BUG=364729

Review URL: https://codereview.chromium.org/261013004