[Android] Ensure MotionEvents are always recycled
There were several cases in ContentViewCore where a MotionEvent was obtained
(copied) for offsetting purposes, but was not recycled because of early function
exit. Fix this by recycling the event in the finally clause. Also avoid copying
touch events if the current touch offset is zero.

Review URL: https://codereview.chromium.org/263833013