Allow overriding Google doodle URL on Android.
This makes the LogoService respect the --google-base-url command line
flag when determining the URL from which to download Google doodles.
This allows easier manual testing of the doodle: if there's a doodle in
France today, set the flag --google-base-url=https://www.google.fr and
the current google in France will be shown.

BUG=369676
NOTRY=true

Review URL: https://codereview.chromium.org/264713013