Skipping WebRTC Firefox inteop test if there's no webcam on the system.
All WebRTC bots used to have software webcams on them, but they don't
anymore since we have transitioned to use internal fake video in Chrome
for almost all the tests. However, Firefox doesn't AFAIK have the any
fake device support which means this test needs a webcam on the
executing system. This patch makes the code look for a webcam and skips
the test if there is no webcam on the executing system.

BUG=364418
TBR=henrika@chromium.org

Review URL: https://codereview.chromium.org/269813003