Highlights of CSS tweaks:
(1) The filename of a <pre> block is now hidden until hover. This fixes the data-filename grey box covering code when the first 1-2 lines of code are long.

(2) In <pre> blocks, <b> and <strike> are more visible

(3) Size of text bumped down slightly to fit more code in but still remain legible. 

(4) Text in <pre> blocks no longer wrap and horizontal scroll is enabled when a line of code is too long.

BUG=none

Review URL: https://codereview.chromium.org/269983002