Fill video frame gaps created after discarding invalid frames.
In MSE, a new segment must start with a key frame.
However, for HLS, this is just a recommendation and some HLS streams
do not comply with that recommendation. The leading non key frames
must be discarded.
However, simply discarding these frames would create a gap
in the video buffer ranges. This gap is now filled with the first
valid key frame.

BUG=None

Review URL: https://codereview.chromium.org/250703005