Fix tab dragging becoming choppy as a result of https://codereview.chromium.org/260263002
It looks like calling Compositor::ScheduleRedrawRect() is unnecessary when the size of the window has not changed. My hope is that X11 Expose events are generated in all of the situations when the window needs to be redrawn (e.g. window moved onscreen, window unminimized)

BUG=370513
TEST=Manual, see bug

Review URL: https://codereview.chromium.org/274483002