[Android] Fix race in setting up accessiblity.
It's possible that the Java side sends us a signal to enable
accessibility before we have a RVH set up, in which case we won't
enable it. Fix that by ensuring accessibility is put into the correct
state when we receive the RENDER_VIEW_HOST_CHANGED notification.

Review URL: https://codereview.chromium.org/273683002