Mojo: Add the ability to allocate space for platform handle info in TransportData.
We don't really need this for POSIX, but we'll need it for Windows, so
all the interfaces will need to take it into account. E.g., if the
source process can duplicate handles into the target, it needs space to
store the new handle value. (If we're using a broker, other data is
needed.)

R=yzshen@chromium.org

Review URL: https://codereview.chromium.org/278513003