Fix crash on fast consecutive item moves in the Windows app list.
This CL fixes an issue where view hierarchy changes in the app list
would invalidate the drag view in the middle of processing a drag update.
This would cause a crash. In order to fix this, check that the drag view
is still valid after the view hierarchy changes.

BUG=371212

Review URL: https://codereview.chromium.org/264413003