Add back the ALERT trick to detect screen readers.
This fires an EVENT_OBJECT_ALERT notification on the legacy hwnd
with a custom id, and if we get a response we enable full web
accessibility.

This code used to be in render_widget_host_view_win. The replacement
code was not working reliably for unknown reasons, and other accessibility
clients were relying on the previous trick - see bugs.

BUG=319523,342319
NOTRY=true
R=ananta@chromium.org, creis@chromium.org

Review URL: https://codereview.chromium.org/272803003