Resubmit: Add real file for AdNetworks DB
This is a resubmit of https://codereview.chromium.org/268063004/.

The previous version was reverted because it caused a memory leak. This was a
problem with RefCounting from ui::ResourceBundle::LoadDataResourceBytes(), which
has been fixed.

Details:
Even though the method returns a non-refcounted ptr to a RefCountedStaticMemory
(which according to the class comment, "the ref counting does not matter" [1]),
the ref-counting _does_ matter. The fix for this went in as part of a refactor
in https://codereview.chromium.org/263953003/.

[1] http://src.chromium.org/viewvc/chrome/trunk/src/base/memory/ref_counted_memory.h?revision=267321 line 44

BUG=357204

TBR=jhawkins@chromium.org (previously approved)
TBR=felt@chromium.org (very minor changes)

Review URL: https://codereview.chromium.org/274563003