CoreAnimation: Ensure frames are acked when tab capture is enabled
If tab capture is enabled, but the window that is being displayed is
occluded, then the quality of the broadcast may be negatively impacted
by throttling of its presentation.

To avoid this, disable throttling the presentation of the windows that
are occluded but may be being captured.

Do not disable throttling for non-occluded windows, because that will
result in the renderer overloading the browser, and frames being dropped
by the browser, resulting in unsmooth animation.

BUG=350410

Review URL: https://codereview.chromium.org/279523003