Adding platform specific variables for MediaCanPlayTypeTest
Vp9 codec is supported after KitKat (API level 19)
Expectation values are modifed in the test case based on the Android sdk version

HLS mime types are supported after Ice-cream sandwich (API level 14)
A check for identifying android version is added which decides whether the mime type should be added in the media_map_ or not.

Theora is not supported on Android. This check is also added in the "IsCodecSupportedOnAndroid".

MediaCanPlayType test are re-enabled on android.

BUG=357665

Review URL: https://codereview.chromium.org/230413003