Fix building android webview on arm64 on mac.
Android uses gcc 4.8 for target on arm64, so -Wno-unused-local-typedefs
gets added to the gcc command line; unfortunately the host toolchain on
mac does not support this flag. We have no way to detect the host
toolchain version in the WebView build because we generate makefiles in
advance instead of on the build machine, so just hardcode an exception
until the instances of this warning can be fixed and this entire
conditional section removed.

BUG=321833

Review URL: https://codereview.chromium.org/277983002