cc: Fix tile manager raster tile iterator perftest.
This test is currently broken, as it doesn't really reset the initial
conditions each run.

NOTE TO PERF SHERIFFS
---------------------

This is very likely to regress all of the
tile_manager_raster_tile_iterator tests. This is expected, as the test
currently doesn't do anything useful. This is a correctness fix for
that test.

R=reveman, enne

Review URL: https://codereview.chromium.org/283463003