[fsp] Add support for opening files.
This patch adds support to opening a file. It is currently no-op, and it is
not yet bound to fileapi.

In order to bound it to fileapi::AsyncFileUtil, we need to implement closing
operation, and reading operations first. Otherwise, creating a snapshot
(required by fileapi::AsyncFileUtil) is not possible.

TEST=unit_tests: *FileSystemProvider*OpenFile*
BUG=248427

Review URL: https://codereview.chromium.org/284443002