gclient: fix --delete_unversioned_trees bug
This occurs for svn checkouts when the DEPS change from:
  third_party/skia/gyp
  third_party/skia/include
  third_party/skia/src
to:
  third_party/skia

First, gclient deletes/moves third_party/skia to make room for the new checkout.
When finished, it notices that the three subdirectories under skia are now
"unversioned" according to .gclient_entries, even though they are part of Skia
repo. On the bots, this causes them to be deleted.

There is a check to determine whether the thought-to-be unversioned directories
are actually subdirectories of another checkout, but it doesn't work in the svn
case because the check assumes that, if they are subdirectories of another
checkout, it must be an svn checkout. scm.GetCheckoutRoot() returns None
because it can't find an svn checkout containing those subdirectories. None is
not in gclient_entries, so the directories get deleted.

BUG=

Review URL: https://codereview.chromium.org/275103009