A couple of incremental improvements for DiscardableMemory.
Unify DCHECKs so that implementation is_locked_ refers uniformly to the
client's perspective, not the underlying implementation's perspective or
the manager's perspective.

Unify Initialize methods to check for "not failed" instead of a
particular form of success.

Don't reset page protection to default after a fresh allocation, which
already delivers pages with default protection.

R=reveman@chromium.org, reveman

Review URL: https://codereview.chromium.org/285493004