Chromoting iOS client
This is a sub-set of:
https://codereview.chromium.org/186733007/

It doesn't include the .xcodeproj files which were causing problems with "git
cl" due to being in the root .gitignore. They won't be necessary after gyp
integration.

This is just a code drop. Nothing is being built at the moment.

NOTRY=true
TBR=sergeyu@chromium.org
BUG=331356

Review URL: https://codereview.chromium.org/278863003