Use LapTimer in rasterize_and_record micro benchmark.
Timing record or rasterize on very small layers can be problematic. The
operations are very short and timing them individually can lead to quantization.
Use a LapTimer to run the operations we're timing for at least a specified
amount of time to avoid this problem.

Moves LapTimer from cc/test to cc/debug so it's accessible outside of testing
code.

BUG=

Review URL: https://codereview.chromium.org/279183002