Fixed PinchViewport getting resized to empty size.
Looks like we forgot to save the visible viewport to the local member
in RenderWidget. This was a problem when DisableAutoResize was called
(such as by the BrowserPluginGuest, as in the bug) since it would
pass the member variable along, resetting the pinch viewport to 0,0.
Since the pinch viewport was recently made to clip its contents, this
resulted in a blank screen.

BUG=371061

Review URL: https://codereview.chromium.org/284883002